Added 64-bit indexed when compiling in 64-bit mode

+++ Binary Bzet/Binary Bzet/bitR.h
@@ -1,14 +1,14 @@
 #ifndef __BITR_H__
 #define __BITR_H__
 
-typedef unsigned int u32;
+typedef size_t uint;
 
 class bitR
 {
 public:
 	bitR(){ m_size = 0; }
 	~bitR(){}
-	void bitR::add(u32 start,u32 end=0, u32 step=0)
+	void bitR::add(uint start,uint end=0, uint step=0)
 	{
 		if(end==0)	end = start+1;
 		if(end > m_size)	m_size = end-1;
@@ -17,21 +17,21 @@ public:
 		m_end.push_back(end);
 		m_step.push_back(step+1);
 	}
-	bool bitR::at(u32 index)
+	bool bitR::at(uint index)
 	{
-		for(u32 i=0; i<m_start.size(); i++)
+		for(uint i=0; i<m_start.size(); i++)
 			if((index >= m_start[i]) && (index < m_end[i]) && ((index-m_start[i]) % m_step[i] == 0))
 				return true;
 		return false;
 	}
-	u32 bitR::size()
+	uint bitR::size()
 	{
 		return m_size;
 	}
 private:
-	u32 m_size;
-	std::vector<u32> m_start;
-	std::vector<u32> m_end;
-	std::vector<u32> m_step;
+	uint m_size;
+	std::vector<uint> m_start;
+	std::vector<uint> m_end;
+	std::vector<uint> m_step;
 };
 #endif
